requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.9.0
attrs==24.2.0
babel==2.17.0
binance-connector==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.2.1
contourpy==1.2.1
cycler==0.12.1
dateparser==1.2.0
dhanhq==1.3
et_xmlfile==2.0.0
fastapi==0.115.14
fonttools==4.53.1
frozenlist==1.4.1
gevent==24.2.1
greenlet==3.0.3
h11==0.16.0
idna==3.7
Jinja2==3.1.6
kiwisolver==1.4.5
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
mt5==0.0.3.1
multidict==6.0.5
mysql-connector==2.2.9
mysql-connector-python==9.3.0
numpy==2.0.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pycryptodome==3.20.0
pyd==0.13.1
pydantic==2.11.7
pydantic_core==2.33.2
pydentic==0.0.1.dev3
pygame==2.6.0
pyparsing==3.1.2
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-stdnum==2.1
pytube==15.0.0
pytz==2024.1
regex==2024.7.24
requests==2.32.3
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.2
tabulate==0.9.0
tkcalendar==1.6.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.35.0
v==1
websocket==0.2.1
websocket-client==1.8.0
websockets==12.0
yarl==1.9.4
yt-dlp==2025.6.30
zope.event==5.0
zope.interface==6.4.post2
